+++ src/picotorrent/ui/dialogs/preferencesdialog.cpp
@@ -5,7 +5,6 @@
 #include <wx/bookctrl.h>
 #include <wx/listbook.h>
 #include <wx/listctrl.h>
-#include <wx/hyperlink.h>
 #include <wx/persist.h>
 #include <wx/persist/toplevel.h>
 #include <wx/simplebook.h>
@@ -57,7 +56,6 @@ PreferencesDialog::PreferencesDialog(wxWindow* parent, std::shared_ptr<pt::Core:
     m_mainSizer->Add(m_book, 1, wxEXPAND);
 
     auto buttonSizer = new wxBoxSizer(wxHORIZONTAL);
-    buttonSizer->Add(new wxHyperlinkCtrl(this, wxID_ANY, i18n("documentation"), "https://docs.picotorrent.org/en/master/configuration.html"), 0, wxALIGN_CENTER_VERTICAL);
     buttonSizer->AddStretchSpacer(1);
     buttonSizer->Add(new wxButton(this, ptID_BTN_RESTORE_DEFAULTS, i18n("restore_defaults")), 0, wxRIGHT, FromDIP(7));
     buttonSizer->Add(new wxButton(this, wxID_OK, i18n("ok")), 0, wxRIGHT, FromDIP(7));